CXP-1101: move audit controller
