+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Audit/Controller/Internal/GetWeeklyAuditAction.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Audit;
+namespace Akeneo\Connectivity\Connection\Infrastructure\Audit\Controller\Internal;
 
 use Akeneo\Connectivity\Connection\Application\Audit\Query\GetPeriodEventCountPerConnectionHandler;
 use Akeneo\Connectivity\Connection\Application\Audit\Query\GetPeriodEventCountPerConnectionQuery;
