+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Audit/Controller/Internal/GetWeeklyErrorAuditAction.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Audit;
+namespace Akeneo\Connectivity\Connection\Infrastructure\Audit\Controller\Internal;
 
 use Akeneo\Connectivity\Connection\Application\Audit\Query\GetPeriodErrorCountPerConnectionHandler;
 use Akeneo\Connectivity\Connection\Application\Audit\Query\GetPeriodErrorCountPerConnectionQuery;
