+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Audit/Controller/Internal/GetErrorCountPerConnectionActionEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\back\tests\EndToEnd\Audit;
+namespace Akeneo\Connectivity\Connection\back\tests\EndToEnd\Audit\Controller\Internal;
 
 use Akeneo\Connectivity\Connection\back\tests\EndToEnd\WebTestCase;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\FlowType;
@@ -18,7 +18,7 @@ use Symfony\Component\HttpFoundation\Response;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class ErrorCountPerConnectionEndToEnd extends WebTestCase
+class GetErrorCountPerConnectionActionEndToEnd extends WebTestCase
 {
     public function test_it_get_error_count_per_connection()
     {
