+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Audit/Controller/Internal/GetWeeklyErrorAuditActionEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\back\tests\EndToEnd\Audit;
+namespace Akeneo\Connectivity\Connection\back\tests\EndToEnd\Audit\Controller\Internal;
 
 use Akeneo\Connectivity\Connection\back\tests\EndToEnd\WebTestCase;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\FlowType;
@@ -16,7 +16,7 @@ use Symfony\Component\HttpFoundation\Response;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class WeeklyErrorAuditEndToEnd extends WebTestCase
+class GetWeeklyErrorAuditActionEndToEnd extends WebTestCase
 {
     public function test_it_get_the_weekly_error_audit(): void
     {
